code/requirements.txt
imitation
pygame
gymnasium
wandb